changed to use floats like in all other scritps

+++ ProjectFiles/Physics/Air.cs
@@ -22,19 +22,19 @@ public static class Air {
         // Calculate air density using the ideal gas law
         return (float)(pressure * molarMassOfDryAir / (idealGasConstant * temperature));
     }
-    public static void GetLocalAirStatistics(float altitude/* h */, out double airDensity, out double temperature) {
-        temperature = seaLevelTemperature - (temperatureLapseRate * altitude);
+    public static void GetLocalAirStatistics(float altitude/* h */, out float airDensity, out float temperature) {
+        temperature = (float)(seaLevelTemperature - (temperatureLapseRate * altitude));
 
         double powX = 1 - (temperatureLapseRate * altitude / seaLevelTemperature);
         double powY = ConstantsManager.gravity * molarMassOfDryAir / (idealGasConstant * temperatureLapseRate);
         double pressure = seaLevelPressure * Math.Pow(powX, powY);
 
         // Calculate air density using the ideal gas law
-        airDensity = (pressure * molarMassOfDryAir / (idealGasConstant * temperature));
+        airDensity = (float)(pressure * molarMassOfDryAir / (idealGasConstant * temperature));
     }
     public static float GetMatchNumber(float velocity, float altitude) {
 
-        GetLocalAirStatistics(altitude, out _, out double temperature);
+        GetLocalAirStatistics(altitude, out _, out float temperature);
 
         float speedOfSound = (float)Mathf.Sqrt((adiabaticIndexForAir * idealGasConstant * temperature) / molarMassOfDryAir);
